fix AuthGroupRepository.getMany
