+++ src/db/sql/index.ts
@@ -1,2 +1,3 @@
 export { SQLs } from './sqls'
 export { SqlSelectBuilder } from './sqlSelectBuilder'
+export { SqlJoinBuilder } from './sqlJoinBuilder'
