+++ src/repository/authGroup/getMany.ts
@@ -1,11 +1,7 @@
 import { AuthGroup, Objects } from '@openforis/arena-core'
-import { BaseProtocol, DB } from '../../db'
-import { TableAuthGroup } from '../../db/table/schemaPublic/authGroup'
-import { SqlSelectBuilder } from '../../db/sql/sqlSelectBuilder'
-import { TableAuthGroupUser } from '../../db/table/schemaPublic/authGroupUser'
-import { SqlJoinBuilder } from '../../db/sql/sqlJoinBuilder'
+import { BaseProtocol, DB, SqlJoinBuilder, SqlSelectBuilder, TableAuthGroup, TableAuthGroupUser } from '../../db'
 
-export const getMany = (options: { userUuid: string }, client: BaseProtocol = DB): Promise<AuthGroup[]> => {
+export const getMany = (options: { userUuid: string }, client: BaseProtocol = DB): Promise<Array<AuthGroup>> => {
   if (!('userUuid' in options)) throw new Error(`missingParams, ${options}`)
   const { userUuid } = options
 
@@ -14,7 +10,7 @@ export const getMany = (options: { userUuid: string }, client: BaseProtocol = DB
 
   const joinClause = new SqlJoinBuilder()
     .join(tableAuthGroup)
-    .on(`${tableAuthGroup.uuid} = ${tableAuthGroupUser.userUuid}`)
+    .on(`${tableAuthGroup.uuid} = ${tableAuthGroupUser.groupUuid}`)
 
   const sql = new SqlSelectBuilder()
     .select(
@@ -29,5 +25,5 @@ export const getMany = (options: { userUuid: string }, client: BaseProtocol = DB
     .where(`${tableAuthGroupUser.userUuid} = $1`)
     .build()
 
-  return client.map(sql, [userUuid], (row) => Objects.camelize(row))
+  return client.map<AuthGroup>(sql, [userUuid], (row) => Objects.camelize(row))
 }
