+++ src/service/user/get.ts
@@ -1,5 +1,27 @@
-import { UserService } from '@openforis/arena-core'
+import { User, UserService, UserStatus } from '@openforis/arena-core'
+import { AuthGroupRepository, UserRepository, UserResetPasswordRepository } from '../../repository'
 
-import { UserRepository } from '../../repository'
+const _initializeUser = async (user: User): Promise<User> => {
+  // Assoc auth groups
+  let userUpdated = {
+    ...user,
+    authGroups: await AuthGroupRepository.getMany({ userUuid: user.uuid }),
+  }
+  if (user.status === UserStatus.INVITED) {
+    const expired = !(await UserResetPasswordRepository.hasValidResetPassword({ userUuid: user.uuid }))
+    userUpdated = {
+      ...userUpdated,
+      invitation: {
+        expired,
+      },
+    }
+  }
 
-export const get: UserService['get'] = UserRepository.get
+  return userUpdated
+}
+
+export const get: UserService['get'] = async (options) => {
+  const user = await UserRepository.get(options)
+  if (user) return _initializeUser(user)
+  return null
+}
